+++ CHANGES.md
@@ -1,6 +1,12 @@
 1.7.0 (in development)
 ======================
 
+dev
+===
+
+- Support for pickling type annotations on Python 3.10 as per PEP 563:
+  https://www.python.org/dev/peps/pep-0563/
+  ([PR #400](https://github.com/cloudpipe/cloudpickle/pull/400))
 
 1.6.0
 =====